Folder refactoring, error handling, and index.html added
A major folder refactoring was done to improve the readability and maintenance of the code. New folders such as router, public, utils, and classes were created to better organize the code. Additionally, an error handling system was implemented in the main app to provide better error messages to the user. Finally, a simple index.html file was added to provide a basic view of the data in the front end.