Added Cart Badge, Render All Products in Cart, and Transaction Alerts
Installed various modules, including socket.io, morgan, and express handlebars. Implemented a live chat feature, product cards, and an add new product form. Added a cart badge that displays the number of products in the cart and a route to render all products in the cart. Implemented transaction alerts using sweet alert to provide feedback to the user.